requirements.txt
python-telegram-bot==21.4
openai==1.38.0
pip install python-dotenv==1.0.1